requirements.txt
altgraph==0.17
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
bcrypt==3.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
docutils==0.17.1
EasyProcess==0.3
entrypoint2==0.2.4
et-xmlfile==1.1.0
ffmpeg==1.4
ffmpeg-python==0.2.0
flake8==3.9.2
fpdf==1.7.2
future==0.18.2
google-auth==1.32.1
google-auth-oauthlib==0.4.4
greenlet==1.1.0
gspread==3.7.0
idna==2.10
isbnlib==3.10.8
isbntools==4.3.27
kaki==0.1.4
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.gstreamer==0.3.2
kivy-deps.sdl2==0.3.1
Kivy-examples==2.0.0
Kivy-Garden==0.1.4
kivy-garden.graph==0.4.0
kivy-garden.xcamera==2020.613
kivy-garden.zbarcam==2020.524
kivymd @ git+https://github.com/HeaTTheatR/KivyMD.git@14b9cc34bd441f91a9ee2c55ee67d3570bb5fcb1
kivymd-extensions.akivymd @ https://github.com/kivymd-extensions/akivymd/archive/main.zip
kivymd-extensions.sweetalert==0.1.5
kiwisolver==1.3.1
ldap3==2.9.1
loguru==0.5.3
matplotlib==3.2.2
mccabe==0.6.1
monotonic==1.6
mss==6.1.0
mysql==0.0.3
mysql-connector-python-rf==2.2.2
mysqlclient==2.0.3
numpy==1.21.0
oauthlib==3.1.1
opencv-python==4.5.2.54
openpyxl==3.0.7
outcome==1.1.0
pandas==1.2.5
pdf2image==1.16.0
pefile==2021.5.24
Pillow==8.3.1
plyer==2.0.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydirectory==0.1.15
pyflakes==2.3.1
Pygments==2.9.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.2
pymongo==3.11.4
pyparsing==2.4.7
PyPDF2==1.26.0
pypiwin32==223
pyscreenshot==3.0
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.19.0
pytz==2021.1
pywin32==301
pywin32-ctypes==0.2.0
pyzbar==0.1.8
pyzbar-x==0.2.1
reportlab==3.5.68
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.0
seaborn==0.11.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.22
SQLAlchemy-Utils==0.37.8
toml==0.10.2
trio==0.19.0
urllib3==1.26.6
watchdog==2.1.3
win32-setctime==1.0.3
xcamera==2020.613
xlrd==2.0.1
zbarcam==2020.524
